+++ serpentarium/__init__.py
@@ -6,6 +6,6 @@ CLEAN_SYS_MODULES = sys.modules.copy()
 
 from . import concurrency
 from .plugin import Plugin, SingleUsePlugin, MultiUsePlugin
-from .abstract_plugin import AbstractPlugin
+from .named_plugin_mixin import NamedPluginMixin
 from .multiprocessing_plugin import MultiprocessingPlugin
 from .plugin_loader import PluginLoader