+++ tests/plugins/run_parameters/plugin.py
@@ -1,8 +1,8 @@
 from typing import Any
 
-from serpentarium import AbstractPlugin
+from serpentarium import MultiUsePlugin, NamedPluginMixin
 
 
-class Plugin(AbstractPlugin):
+class Plugin(NamedPluginMixin, MultiUsePlugin):
     def run(self, my_param: Any):  # type: ignore[override]
         return my_param