+++ CHANGELOG.md
@@ -10,6 +10,9 @@ the [PEP 440 version scheme](https://peps.python.org/pep-0440/#version-scheme).
 ### Added
 - py.typed
 
+### Changed
+- AbstractPlugin to NamedPluginMixin
+
 ### Fixed
 - The Alpha status classifier